Missed 3 compiler warnings under VS2005
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@3815 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/ui/Windows/PasswordSubsetDlg.cpp
@@ -136,7 +136,7 @@ LRESULT CPasswordSubsetDlg::OnDisplayStatus(WPARAM /* wParam */, LPARAM /* lPara
       continue;
 
     int ipos = _wtoi(resToken);
-    if (ipos > ipwlengh || ipos == 0) {
+    if (ipos > (int)ipwlengh || ipos == 0) {
       if (ipos != 0)
         m_warningmsg.Format(IDS_SUBSETINDEXTOOBIG,ipwlengh);
       else
